requirements.txt
Flask
Gunicorn
cloudinary
python-dotenv